bot/requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=1.0.0
